Added functions that generate dummy signatures
We want to unblock people when our service isn't working and take some
pressure off keeping the damn thing up.

This PR hard codes in a private key with a value of 0 and a nonce with a
value of 1 and does accurate key derivation on it. Provided you fetch
the dummy root key, client code should behave consistently. The only
small difference is I don't include episilon or a changing nonce meaning
that signatures with an identical predecessor, derivation and payload
will be identical.

This currently doesn't build on WASM and I suspect my moving around of
dependencies will have bloated the size of the contract.