Make sure when we fail in the busy loop we cleanup (#586)
* Make sure when we fail in the busy loop we cleanup

We got reports of users getting responses to sign requests on testnet of
"Too many pending requests. Please, try again later." when the signing
contract was under no load.

I believe this was caused by someone sending a message that failed, the
contract cleaning up the request but then immediately panicking. This
caused the state cleanup to roll back and the system keeping a limit on
the total number of requests to eventually run out of space.

This PR fixes that by panicking on a seperate call and allowing the
cleanup to persist.