feat: poc of a request queue in redis
This commit shows how a redis list can be used to decouple
the sign request producers (indexers) from the rest of the MPC node.

This is essentially just ripping out a mpsc::channel() with a redis FIFO queue.
Most code changes are just replacing the corresponding sender / receiver to
a redis pool.

The biggest change in this is that we have to serialize `SignRequest` to queue it.
This caused a conflict with the time_added field on the request.
Using `Instant` only works for relative timing within a single process and it is
therefore not designed to be serialized. Replacing it with `SystemTime` works
opens up problems with system time changes, in which case the `.elapsed()` time
could fail to work or return invalid data. I think this is fine, as we only use the time
for metrics. Thus, I use `unwrap_or_default` to convert failures in time measuring
to 0 ms duration.