refactor: delete unused `is_stable` code
`is_stable` has been set to a constant of `true` since #241

The `is_stable` flag was useful when we only looked at the near indexer.
It checked of the indexer data is behind.
But now with many chains integrated, we want to move away from
chain-specific checks. Hence, is_stable should go away entirely.

This PR removes dead code that was still around from the old is_stable check
and  renames some field to make it easier to understand what actually happens.

Specifically, replace `NodeStatus::Active(IsStable=true)` with `NodeStatus::Active`.

For the case where the peer node is online but it still is in a protocol state
where it is not ready to participate, we now use `NodeStatus::Inactive`.
(Before, we also used to `is_stable` flag for that, although it doesn't really have anything to do with stability.)

These changes are in preparation for state sync to be used for deciding which
nodes are indeed ready to participate in a protocol invocation.
For that, we want to add a check for a node being in-sync.
Which will touch the same code, so it's a good timing to refactor it a bit first.