feat: only set nodes as active after state sync
This introduces NodeState::Syncing per connection on the mesh level.

The SyncTask will take peers in this state and prioritize syncing them.
After it was synced, this is reported back to the mesh using a message channel.

This PR tries to introduce this new state without changing too much architecture and code. I'm not super clear about it yet, how exactly the code
would be refactored best. For now, this seems to be good enough to me
but I am very much open to ideas.