Fix Solana conflicts (#323)
* performance metrics: fix triple metrics (#303)

* fix triple metrics: modify buckets, and lock the poked_latest for triple generator

---------

Co-authored-by: Xiangyi Zheng <xyzheng@uhicago.edu>

* perf metrics: increase the upper range for accrued delay and triple pokes (#309)

* perf metrics: increase the upper range for accrued delay

* up range for triple pokes

---------

Co-authored-by: Xiangyi Zheng <xyzheng@uhicago.edu>

* feat: sync remove storage (#304)

* Added triple_owner to triple storage

* Added benchmarks for storage

* Move storage tests over to store.rs

* Added triple reservation

* Added remove_oudated and update storage.clear

* Added storage features to presignature storage

* Make taking mine remove from owner set

* signature test working again

* Cleanup sync

* Made sync use mine | reserved

* Cleanup triple outcome

* Added back threshold check to signature

* eth indexer: indexer delay metric (#312)

* eth indexer: indexer delay metric


---------

Co-authored-by: Xiangyi Zheng <xyzheng@uhicago.edu>

* fix: race condition on sync and unable to do presignature reservation on error (#311)

* Made PresignatureSlot unreserve on drop

* Make take_two reserve a slot so owner doesn't unsync others

* Make presig.take reserve a slot so owner doesn't unsync others

* Added test

* Make TripleSlot unreserve on err

* Clippy & add additional test

* Fix clippy

* feat: Helios indexer (#306)

* helios: use helios for eth indexer

* use alloy instead of web3 whenever possible, except web3::signing::recover

---------

Co-authored-by: Xiangyi Zheng <xyzheng@uhicago.edu>

---------

Co-authored-by: Xiangyi Zheng <xyzheng@uhicago.edu>

* Respect CARGO_TARGET_DIR in build script (#315)

Some devs (like myself) like placing the target dir for multiple
projects in a common shared place to keep the size of all locals builds
smaller. `CARGO_TARGET_DIR` exists for this.

The integration tests should ideally check this env variable rather than
assuming the directory is named exactly "target".

* refactor: delete unused `is_stable` code (#317)

`is_stable` has been set to a constant of `true` since #241

The `is_stable` flag was useful when we only looked at the near indexer.
It checked of the indexer data is behind.
But now with many chains integrated, we want to move away from
chain-specific checks. Hence, is_stable should go away entirely.

This PR removes dead code that was still around from the old is_stable check
and  renames some field to make it easier to understand what actually happens.

Specifically, replace `NodeStatus::Active(IsStable=true)` with `NodeStatus::Active`.

For the case where the peer node is online but it still is in a protocol state
where it is not ready to participate, we now use `NodeStatus::Inactive`.
(Before, we also used to `is_stable` flag for that, although it doesn't really have anything to do with stability.)

These changes are in preparation for state sync to be used for deciding which
nodes are indeed ready to participate in a protocol invocation.
For that, we want to add a check for a node being in-sync.
Which will touch the same code, so it's a good timing to refactor it a bit first.

* feat: use generic owner set over mine set (#307)

* Added triple_owner to triple storage

* Added benchmarks for storage

* Move storage tests over to store.rs

* Added triple reservation

* Added remove_oudated and update storage.clear

* Added storage features to presignature storage

* Make taking mine remove from owner set

* signature test working again

* Cleanup sync

* Made sync use mine | reserved

* Cleanup triple outcome

* Added back threshold check to signature

* Remove mine set for storage

* Remove actual mine key

* Bump storage version

* Remove triple clone

* Make storage not return StoreResult

* Fix test

* Remove _clear

* Remove sync integration test for now

* add artifacts required to compile the project to the repo (#316)

We statically include the artifact

`contract-eth/artifacts/contracts/ChainSignatures.sol/ChainSignatures.json`

at chain-signatures/node/src/rpc.rs:L381.

If we want to keep it that way, I think it would make sense to check it into the
source repository, since currently the project simply doesn't compile after a
fresh checkout.

* Reorganize sign request participants on failure (#321)

* FEAT: initial load test specifications (#320)

* initial load test specifications

* separated environments into groups, added test for develop and main branches independanatly

* added tests according to discussed plan, updated workflows to be ran ad-hoc

* reduced duplicate files, passed env vars for contract addresses

* updated test file names

* increased total time to 2 hours, reduced response time to 1.5s

* fmt

* clippy

* use from_config for solana

* remove macro

* add macro back

* fix warnings and failing unit tests

* fix default args

* fix default args

* use Osman's latest contract and he deployed version on solana devnet

* debug prints

* fmt

* new solana contract address

* clippy

---------

Co-authored-by: Xiangyi Zheng <xyzheng@uchicago.edu>
Co-authored-by: Xiangyi Zheng <xyzheng@uhicago.edu>
Co-authored-by: Phuong Nguyen <ChaoticTempest@users.noreply.github.com>
Co-authored-by: Jakob Meier <inbox@jakobmeier.ch>
Co-authored-by: Kody Maus <kody@sig.network>
Co-authored-by: Bo Yao <by677@nyu.edu>