remove NearClient from MpcSignProtocol
Removing the dependency makes the code more modular and testable.

The NearClient is used to read the governance smart contract on-chain state. Currently, this is on near but it may change in the future.

Aside from what we call the ConsensusProtocol, the MPC protocols don't need to know about the governance. Thus, I added a new parameter to the consensus protocol. Now we can test other protocol without a NearClient.

Finally,  I replace the NearClient parameter with a generic type parameter, to allow mocking it in tests.