component testing the MPC
Testing framework for testing the MPC component without a real network.

These tests create a test fixture with an in-memory node-to-node network
as well as a in-memory shared governance contract proxy. With that, MPC
testing should be almost entirely deterministic.

The only external dependency a test fixture still needs is a redis
container.

This separation of the MPC component from the rest of the system is a
first move in a general new direction of testing. The long-term goal
is to replace the few integration tests that run in CI today with a much
larger number of tests that are more light-weight and deerministic.

While the component tests for MPC are still few and simple, the setup
lays down a foundation for many more tests. The new framework gives
total control for the order in which messages are received, including
malicious messages. Adding such tests will come in a separate PR to
not make this PR even larger.

(key sharing, triple generation, presignature genreation, signature)