Mpc testing v2.1 (#487)
* remove NearClient from MpcSignProtocol

Removing the dependency makes the code more modular and testable.

The NearClient is used to read the governance smart contract on-chain state. Currently, this is on near but it may change in the future.

Aside from what we call the ConsensusProtocol, the MPC protocols don't need to know about the governance. Thus, I added a new parameter to the consensus protocol. Now we can test other protocol without a NearClient.

Finally,  I replace the NearClient parameter with a generic type parameter, to allow mocking it in tests.

* component testing the MPC

Testing framework for testing the MPC component without a real network.

These tests create a test fixture with an in-memory node-to-node network
as well as a in-memory shared governance contract proxy. With that, MPC
testing should be almost entirely deterministic.

The only external dependency a test fixture still needs is a redis
container.

This separation of the MPC component from the rest of the system is a
first move in a general new direction of testing. The long-term goal
is to replace the few integration tests that run in CI today with a much
larger number of tests that are more light-weight and deerministic.

While the component tests for MPC are still few and simple, the setup
lays down a foundation for many more tests. The new framework gives
total control for the order in which messages are received, including
malicious messages. Adding such tests will come in a separate PR to
not make this PR even larger.

(key sharing, triple generation, presignature genreation, signature)

* increase timeout in test

* optimized build for curve25519-dalek in tests

This improves the time for `test_basic_generate_triples` from ~320s
down to ~16s on my machine.