test: Message counting tests
Add tests checking that T / P / S protocols sent the expected amount
of messages.

As noted in the comment, we may not necessarily care about the
exact number. But we should understand roughly what they are
and run tests to ensure it doesn't change unnoticed.

To implement this test, I also made the testing framework more
flexible in how it observed outputs. We can now check specific
counts for each protocol invocation.

The printed debug summary per test looks a bit like this:

```
           1    Forwarded Signature(15446211684585113244)

           1    Forwarded Posit(0)
           1    Forwarded Signature(15446211684585113244)

           2    Forwarded Posit(0)
           1    Forwarded Signature(15446211684585113244)

           1    Forwarded Signature(15446211684585113244)
           2    Forwarded Posit(0)

           1    Forwarded Posit(0)
           1    Forwarded Signature(15446211684585113244)

           1    Forwarded Signature(15446211684585113244)
```