test: interrupt signing
This component test simulates when a node that accepted a posit
fails and loses their presignatures.

The other nodes should eventually detect that the signature is
not produced and start a new round of posits.

Note that this is not BFT, yet. The failing node behaves quite nice,
it will reject future posit rounds. So, the other nodes should have
no problem establishing a new set of participants that works.