fix: align posit rounds
Include the current round in posit messages and avoid mixing messages meant for
different rounds.

When a timeout occurs, use the largest round
that peer messages used, in order to catch up
if other nodes are more than one round ahead.

Lastly, keep a buffer for messages that are
sent to future rounds. Nodes with a deliberator
might otherwise miss the Propose message
if they enter the new round after the proposer.